+++ src/storage/v2/CMakeLists.txt
@@ -39,6 +39,7 @@ set(storage_v2_src_files
     replication/rpc.cpp
     replication/replication.cpp
     inmemory/replication/replication_server.cpp
+    replication/InMemoryReplicationClient.cpp
 )
 
 #######################
